requirements.txt
# Core dependencies
python-dotenv==1.0.0
anthropic==0.39.0

# LangChain & LangGraph
langchain==0.3.7
langchain-core==0.3.15
langchain-anthropic==0.2.4
langgraph==0.2.45

# MCP (Model Context Protocol)
mcp==1.1.2
fastmcp==0.2.2

# Vector store & embeddings
faiss-cpu==1.8.0
sentence-transformers==2.2.2

# Document processing
pypdf==3.17.0

# Web search
tavily-python==0.3.0

# Web interface
streamlit==1.40.0

# Data visualization
matplotlib==3.8.0
pandas==2.1.0
numpy==1.24.0

# Image processing
pillow==10.4.0

# Async support
nest-asyncio==1.6.0
